Support multiple blueprints - Review fixes
